requirements.txt
# ===================================
# FACTURIA 2.0 - Dependencias Python
# ===================================

# === Core ===
python-dotenv==1.0.0
pydantic==2.5.0

# === Email ===
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# === PDF Processing ===
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0

# === AI / Google Gemini ===
google-generativeai==0.3.1

# === Database ===
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # Para PostgreSQL (opcional)

# === Dashboard ===
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.3
numpy==1.26.2

# === Utilities ===
python-dateutil==2.8.2
schedule==1.2.0
loguru==0.7.2

# === Development ===
pytest==7.4.3
black==23.11.0
flake8==6.1.0
